requirements.txt
scrapy
django-tables2
